refactor(agent): split macOS collector modules
